lua: setup loaders before loading built-in modules
The built-in modules overriding functionality will be implemented as a
Lua loader. It must be in effect, when built-in modules are loading, so
setup the loaders earlier.

This commit doesn't change any user visible behavior, but it marks a
minor problem with a filename assigned to the loaded Lua code (seen in
error messages and `debug.getinfo()`) to fix it later.

Part of #7774

NO_TEST=no user visible changes
NO_CHANGELOG=see NO_TEST
NO_DOC=see NO_TEST