memtx: validate only cmp_def parts in pagination
Currently, we check that tuple passed as position fits space format.
However, for pagination purposes, it's enough to validate only parts of
tuple that are used in cmp_def. And, it allows not to use space format -
we will need it in read view. So the patch replaces tuple validation with
validation of its cmp_def parts only.

Closes #8511
Part of tarantool/tarantool-ee#285

NO_DOC=bugfix