test: rewrite test for gh-5983 with luatest
As a result of the patch, all required files are moved to app-luatest
subdirectory. The test itself is rewritten via <justrun> and <treegen>
helpers to make it more clear and easy to maintain.

Since luatest provides <skip_if> helper, .skipcond file is not necessary
anymore and is removed.

Needed for #8252

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring