log: fix flight-recording of json log messages
When log format is JSON and a Lua table is written to the log, such
messages are saved by the flight recorder as a "json" string. Fix it.

Part of tarantool/tarantool-ee#325

NO_DOC=bugfix
NO_TEST=will be added to EE, because there are no flightrec in CE
NO_CHANGELOG=will be added to EE, because there are no flightrec in CE