sql: drop ASC/DESC in index creation statements
Prior to this patch, the SQL parser supported the sort order in the
index definition during index creation. However, all of our indexes are
ascending until #3243 is implemented, so the specified sort order was
simply ignored. To avoid a drastic change in behavior due to the
implementation of multidirectional indexes, we have decided to drop the
sort order from the index definition until #3243 is implemented.

Part of #8100

@TarantoolBot document
Title: ASC/DESC during index creation

Now the SQL parser does not support ASC/DESC when creating an index.
This used to be supported by the parser, but never worked.

An example of statements that are no longer supported:
```
CREATE TABLE t(i INT PRIMARY KEY ASC);
CREATE TABLE t(i INT PRIMARY KEY, a STRING, UNIQUE(a DESC));
CREATE INDEX i1 ON t(a ASC);
```