box: refactor `check_select_safety` and add it to `box.internal`
Refactor logging of dangerous `select` call out of `check_select_safety`
and add it to `box.internal` so that it can be reused for read views.

Change all occurrences of 'dangerous' to 'long' to be consistent with the
log message.

Needed for tarantool/tarantool-ee#211

NO_CHANGELOG=<refactoring>
NO_DOC=<refactoring>
NO_TEST=<refactoring>