test: add a generator for a tree of Lua files
This commit adds a generator of Lua file tree, which generates scripts
using provided templates and filenames. This approach is used in
preload_test.lua and I'm going to use it in a future overload_test.lua.

The original code from preload_test.lua remain almost same. It is
arranged to be a module and adopted for convenient use from a luatest
based test. Also, it gains ability to provide several templates for Lua
files instead of just one for all the files.

Part of #7774

NO_CHANGELOG=a testing helper is added
NO_DOC=see NO_CHANGELOG