lua: add internal modules initialization status
The flag indicates, whether the loading of built-in modules is finished.
May be useful for override modules to determine, whether it is required
from inside tarantool or from application's code.

Follows up #7774

NO_DOC=It is not intended to be officially supported. I would prefer to
       document ability of loading a replacement module itself, but I
       wouldn't document how to write such a module.
NO_CHANGELOG=see NO_DOC