debugger: match paths suffixes in breakpoints
We used to "normalize" breakpoint filename paths
down to their basic name, which worked fine for
unique file names, but is getting annoying when
you need to debug a code in one of many `init.lua`
modules, i.e.

    break init.lua:5

Will stop at _every_ `init.lua` file we will run
at the debugging session.

So we add machinery for partial paths lookup, to
make possible to set breakpoints using (unique-enough)
path suffixes, e.g. if we have multiple `init.lua` in
the application, i.e.
    ./init.lua
    ./A/init.lua
    ./B/init.lua

Then we could use syntax:

    break A/init.lua:5

To activate debugger breakpoint only in the particular
module, and not trigger it elsewhere.

Current suffix trees limitations and peculiarities
--------------------------------------------------

Please keep in mind that suffix-tree algorithm, used for
partial paths lookup, uses `$ref` and `$f` node names
for their own purposes. That means that it would not handle
well breakpoints with paths containing `$f` and `$ref`.
That's not a big problem given that majority of filesystems
we are running on do not usually allow `$` as part of
file name.

Also, due to suffix tree lookup behavior, if we would
activate multiple breakpoints:
    break init.lua:10
    break A/init.lua:10

Then _least specific_ `init.lua:10` breakpoint will trigger.
From users' prospective it makes no much difference, as we
need to stop in debugger shell in this line, but this is
the way how suffix tree is working.

Dot files treatment
-------------------

There is a special mechanism for `.` and `..` treatment
if they are at the header of a path provided. I.e.

    break ./main.lua:10
    break ../a/b/c/least.lua:15

For these cases `.` and `..` will be converted to the
full path to the given file location.

NO_TEST=hard to implement tests. Postponed till step.5

@TarantoolBot document
Title: Breakpoints in console debugger for Lua

Match partial path patterns in breakpoints
==========================================

Please see `third_party/lua/README-luadebug.md` for
a fuller description of partial path syntax one could use
in a breakpoints definition.