sql: drop struct sql_stmt
This structure is used as another name for the struct Vdbe. Because of
this, there is some confusion about when one structure or another is
used. To avoid such confusion, we must remove the sql_stmt structure.

In addition, this patch removes several unused functions or non-existent
function definitions.

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring