prbuf: fix losing records on prepare on wrap
There is an issue in prbuf_prepare on branch when we need to
wrap and allocate space for the record starting from the beginning of
the data area.

In the process we set end offset to 0. This means empty buffer. So if
application is crashed after such prepare but before commit buffer will
have no records. This is a bug.

But actually we don't have to reset end to 0 unless we preparing for
very big record which will occupy almost all buffer. In this case there
is really no records left after preparation. Otherwise if we don't
change end offset buffer will be valid and records will be there.

Part of https://github.com/tarantool/tarantool-ee/issues/319

NO_DOC=bugfix
NO_CHANGELOG=minor