test: fix fail of gh_4669_applier_reconnect_test
In the test we start replicas only with master in box.cfg.replication.
We cannot use bootstrap_strategy = 'auto' mode, which is default, as
it properly works only when all participants of the cluster are listed
in replication parameter. Sometimes, when one replica connects to the
master, the other one has already successfully joined, so the first
replica sees in ballot, that it doesn't have all nodes from cluster
in box.cfg.replication and fails to start.

Let's use 'legacy' bootstrap strategy for now.

Closes tarantool/tarantool-qa#310

NO_DOC=test-fix
NO_CHANGELOG=test-fix