sql: add return value check for sql_key_info_new()
Prior to this patch, the return value of sql_key_info_new() in
multiSelectOrderBy() was not checked. This patch fixes that.

This and similar issues have been fixed in version 2.11 and later with
the commit "sql: refactor memory allocation system".

Closes tarantool/security#117

NO_DOC=bugfix
NO_TEST=hard to reproduce the bug
NO_CHANGELOG=bugfix for unlikely bug